# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VC Brain Intelligence API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/candidates": {
      "get": {
        "tags": [
          "candidates"
        ],
        "summary": "Candidates",
        "operationId": "candidates_api_v1_candidates_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "outbound_detector",
                    "inbound_application"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "candidate",
                    "screened",
                    "memo_ready"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "score",
                "momentum"
              ],
              "type": "string",
              "default": "score",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidates/{login}": {
      "get": {
        "tags": [
          "candidates"
        ],
        "summary": "Candidate",
        "operationId": "candidate_api_v1_candidates__login__get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidates/{login}/evidence": {
      "get": {
        "tags": [
          "candidates"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_v1_candidates__login__evidence_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidates/{login}/memo": {
      "get": {
        "tags": [
          "candidates"
        ],
        "summary": "Memo",
        "operationId": "memo_api_v1_candidates__login__memo_get",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Memo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claim_id}": {
      "get": {
        "tags": [
          "claims"
        ],
        "summary": "Claim",
        "operationId": "claim_api_v1_claims__claim_id__get",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Types"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deepdive": {
      "post": {
        "tags": [
          "deepdive"
        ],
        "summary": "Start Deepdive",
        "operationId": "start_deepdive_api_v1_deepdive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeepDiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepDiveAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deepdive/runs": {
      "get": {
        "tags": [
          "deepdive"
        ],
        "summary": "List Deepdives",
        "operationId": "list_deepdives_api_v1_deepdive_runs_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeepDiveRunSummary"
                  },
                  "title": "Response List Deepdives Api V1 Deepdive Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deepdive/runs/{run_id}": {
      "get": {
        "tags": [
          "deepdive"
        ],
        "summary": "Get Deepdive",
        "operationId": "get_deepdive_api_v1_deepdive_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepDiveRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deepdive/runs/{run_id}/stream": {
      "get": {
        "tags": [
          "deepdive"
        ],
        "summary": "Stream Deepdive",
        "operationId": "stream_deepdive_api_v1_deepdive_runs__run_id__stream_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "gh_login": {
            "type": "string",
            "title": "Gh Login"
          },
          "founder_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Founder Name"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "source": {
            "type": "string",
            "enum": [
              "outbound_detector",
              "inbound_application"
            ],
            "title": "Source"
          },
          "current_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Current Score"
          },
          "score_percentile": {
            "type": "number",
            "title": "Score Percentile"
          },
          "momentum_3mo": {
            "type": "number",
            "title": "Momentum 3Mo"
          },
          "first_detection_month": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Detection Month"
          },
          "status": {
            "type": "string",
            "enum": [
              "candidate",
              "screened",
              "memo_ready"
            ],
            "title": "Status"
          },
          "has_memo": {
            "type": "boolean",
            "title": "Has Memo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gh_login",
          "founder_name",
          "company",
          "source",
          "current_score",
          "score_percentile",
          "momentum_3mo",
          "first_detection_month",
          "status",
          "has_memo"
        ],
        "title": "Candidate"
      },
      "CandidateDetail": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "trajectory": {
            "items": {
              "$ref": "#/components/schemas/TrajectoryPoint"
            },
            "type": "array",
            "title": "Trajectory"
          },
          "three_axis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThreeAxis"
              },
              {
                "type": "null"
              }
            ]
          },
          "memo_available": {
            "type": "boolean",
            "title": "Memo Available"
          },
          "evidence_counts_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Evidence Counts By Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate",
          "trajectory",
          "three_axis",
          "memo_available",
          "evidence_counts_by_type"
        ],
        "title": "CandidateDetail"
      },
      "CandidatePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "CandidatePage"
      },
...
      "ClaimResponse": {
        "properties": {
          "claim": {
            "$ref": "#/components/schemas/Claim"
          },
          "resolved_evidence": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvidenceEvent"
                },
                {
                  "$ref": "#/components/schemas/UrlEvidence"
                }
              ]
            },
            "type": "array",
            "title": "Resolved Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "claim",
          "resolved_evidence"
        ],
        "title": "ClaimResponse"
      },
      "DeepDiveAccepted": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "DeepDiveAccepted"
      },
      "DeepDiveRequest": {
        "properties": {
          "entity_type": {
            "type": "string",
            "const": "founder",
            "title": "Entity Type",
            "default": "founder"
          },
          "entity_id": {
            "type": "string",
            "minLength": 1,
            "title": "Entity Id"
          },
          "dimensions": {
            "items": {
              "type": "string",
              "enum": [
                "founder",
                "market",
                "idea_vs_market"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Dimensions"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "live",
                  "replay"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entity_id"
        ],
        "title": "DeepDiveRequest"
      },
      "DeepDiveRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "provenance": {
            "type": "string",
            "const": "live",
            "title": "Provenance",
            "default": "live"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RunStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/LiveEvidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "claims": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "object",
            "title": "Claims"
          },
          "dimension_notes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Axis"
            },
            "propertyNames": {
              "enum": [
                "founder",
                "market",
                "idea_vs_market"
              ]
            },
            "type": "object",
            "title": "Dimension Notes"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gaps"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "token_usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Token Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "entity_id",
          "started_at",
          "finished_at",
          "steps",
          "evidence",
          "claims",
          "dimension_notes",
          "gaps",
          "model",
          "token_usage"
        ],
        "title": "DeepDiveRun"
      },
      "DeepDiveRunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "OK",
                  "INSUFFICIENT_EVIDENCE",
                  "ERROR"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "claim_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Claim Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "entity_id",
          "started_at",
          "finished_at",
          "outcome",
          "claim_count"
        ],
        "title": "DeepDiveRunSummary"
      },
...
      "EvidencePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceEvent"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "EvidencePage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Memo": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "founder_logins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Founder Logins"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "sections": {
            "$ref": "#/components/schemas/MemoSections"
          },
          "claims": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "object",
            "title": "Claims"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gaps"
          },
          "three_axis": {
            "$ref": "#/components/schemas/ThreeAxis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "company",
          "founder_logins",
          "generated_at",
          "sections",
          "claims",
          "gaps",
          "three_axis"
        ],
        "title": "Memo"
      },
...
      "SearchResponse": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/SearchGroup"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "groups"
        ],
        "title": "SearchResponse"
      },
...
      "UrlEvidence": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "UrlEvidence"
      },
...
    }
  }
}